SIGIR_eCommerce_Challenge_2021/task2_purchase_prediction/code/xgboost/sigir_ecom_challenge_code/submission/requirements.txt
boto3==1.15.8
python-dotenv==0.13.0
gensim==4.0.1